requirements.txt
Django==2.2.4
django-extensions==2.2.1
django-rest-framework==0.1.0
django-cors-headers==3.0.2
mysqlclient==1.4.2.post1
itsdangerous==1.1.0
PyJWT==1.7.1
